requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
httpx
python-telegram-bot==20.7
ddgs
python-multipart
pymupdf
psycopg2-binary
qdrant-client
fastembed